requirements.txt
flask
flask-sqlalchemy
flask-mail
psycopg2-binary
python-dotenv
gunicorn
